// front/src/enums.ts
export  enum ModalWindowEnum{
    MODAL_WINDOW_NOT_VALID = -1,
    MODAL_WINDOW_CREATED_CHANNEL = 0,
    MODAL_WINDOW_FILE_ADDED = 1,
    MODAL_WINDOW_UPDATE_USER_NAME = 2,
    MODAL_WINDOW_UPDATE_PASSWORD = 3, 
    MODAL_WINDOW_UPDATE_EMAIL = 4,
    MODAL_WINDOW_DELETE_PROFILE = 5,
    MODAL_WINDOW_FRIEND_INVITE_CHAT = 6,
}